refactor: split off PaginationHeader into separate file
